add delete confirmation

+++ js/controller/delete-jurusan.js
@@ -2,16 +2,19 @@ function deleteData(IDHAPUS) {
   var jurusanId = IDHAPUS;
   var target_url = "https://ws-dito.herokuapp.com/delete-jurusan/" + jurusanId;
 
-  var requestOptions = {
-    method: "DELETE",
-    redirect: "follow",
-  };
+  var confirmation = window.confirm("Apakah Anda yakin ingin menghapus data ini?");
+  if (confirmation) {
+    var requestOptions = {
+      method: "DELETE",
+      redirect: "follow",
+    };
 
-  fetch(target_url, requestOptions)
-    .then((response) => response.json())
-    .then((result) => {
-      alert(result.message);
-      location.reload();
-    })
-    .catch((error) => console.log("Error:", error));
+    fetch(target_url, requestOptions)
+      .then((response) => response.json())
+      .then((result) => {
+        window.alert(result.message);
+        location.reload();
+      })
+      .catch((error) => console.log("Error:", error));
+  }
 }
